benchmarks/requirements.txt
google-cloud-bigquery==3.26.0
